Version 2
for this version 2, I import the art module and print the art which is Caesar Cipher. I also combine 2 function encrypt and decrypt in one function called caesar. Instead of just encode and decode only one single word like the first version, in this version, I modify a little bit so the program can work on the whole passage. 
I also add the while loop so that the program only stop when the user type No. Otherwise, the program will continue, not just execute one time